🎨: デザインシステム連携のため余白基本パターンの記述構成を見直す
